refactor(general): Adjust some imports/exports, move isValidElement
