+++ packages/inferno-create-class/src/createClass.ts
@@ -1,4 +1,4 @@
-import { isFunction, isNullOrUndef, isUndefined } from 'inferno-helpers/lib/index';
+import { isFunction, isNullOrUndef, isUndefined } from 'inferno-helpers';
 
 import Component from 'inferno-component';
 import { ComponentSpec } from '../../inferno-component/src/es2015';
