+++ packages/inferno-create-element/src/index.ts
@@ -1,3 +1,8 @@
 import createElement from './createElement';
+import isValidElement from './isValidElement';
 
 export default createElement;
+
+export {
+  isValidElement,
+};
