+++ packages/inferno-router/src/match.ts
@@ -1,6 +1,6 @@
 import Inferno from 'inferno';
 import pathToRegExp from 'path-to-regexp-es6';
-import { isArray, toArray } from 'inferno-helpers/lib/index';
+import { isArray, toArray } from 'inferno-helpers';
 import { decode, flatten, getURLString, isEmpty, mapSearchParams, pathRankSort, toPartialURL } from './utils';
 
 const cache: Map<string, IMatchRegex> = new Map();
