+++ packages/inferno-test-utils/src/testUtils.ts
@@ -1,7 +1,7 @@
 import { render } from '../../inferno/src/rendering';
 import createElement from '../../inferno-create-element/src/createElement';
-import isValidElement from './isValidElement';
-import { isArray, throwError, toArray } from 'inferno-helpers/lib/index';
+import isValidElement from '../../inferno-create-element/src/isValidElement';
+import { isArray, throwError, toArray } from 'inferno-helpers';
 
 export function renderIntoDocument(element: VNode): VNode {
 	const div = document.createElement('div');
