+++ packages/inferno/inferno.d.ts
@@ -129,6 +129,8 @@ declare module 'inferno-create-class' {
 
 declare module 'inferno-create-element' {
 	export default function createElement(name: any, props?: any, ...children): VNode;
+
+	export function isValidElement(obj: VNode): boolean;  // TODO maybe move to helpers? ask core devs
 }
 
 declare module 'inferno-hyperscript' {
