+++ packages/inferno/src/hydration.ts
@@ -11,7 +11,7 @@ import {
 	isStringOrNumber,
 	isUndefined,
 	throwError,
-} from 'inferno-helpers/lib/index';
+} from 'inferno-helpers';
 import { svgNS } from './constants';
 import Lifecycle from './lifecycle';
 import {
