+++ packages/inferno/src/rendering.ts
@@ -5,7 +5,7 @@ import {
 	isNullOrUndef,
 	NO_OP,
 	throwError,
-} from 'inferno-helpers/lib/index';
+} from 'inferno-helpers';
 
 import options from './core/options';
 import { cloneVNode } from './core/VNodes';
